HumanBehaviour/requirements.txt
absl-py==1.0.0
apparmor==3.0.4
appdirs==1.4.4
asttokens==2.0.5
astunparse==1.6.3
auto-cpufreq==1.9.3
backcall==0.2.0
Beaker==1.11.0
beautifulsoup4==4.10.0
Brlapi==0.8.3
btrfsutil==5.17
CacheControl==0.12.6
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
click==8.1.3
colorama==0.4.4
configobj==5.1.0.dev0
contextlib2==0.6.0.post1
cryptography==36.0.2
cupshelpers==1.0
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
distlib==0.3.4
distro==1.7.0
dnspython==2.2.1
docker-pycreds==0.4.0
docopt==0.6.2
entrypoints==0.4
envycontrol==2.1.1
executing==0.8.3
flatbuffers==2.0
fonttools==4.33.3
gast==0.5.3
gbinder-python==1.0.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.46.0
h5py==3.6.0
html5lib==1.1
httplib2==0.20.3
idna==3.3
importlib-metadata==4.8.1
injector==0.19.0
ipykernel==6.13.0
ipython==8.3.0
jedi==0.18.1
jupyter-client==7.3.0
jupyter-core==4.10.0
keras==2.8.0
Keras-Preprocessing==1.1.2
keyutils==0.6
kiwisolver==1.3.2
LibAppArmor==3.0.4
libclang==14.0.1
libtorrent==2.0.6
louis==3.21.0
lxml==4.8.0
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
menulibre==2.2.3
more-itertools==8.10.0
msgpack==1.0.3
mugshot==0.4.3
nest-asyncio==1.5.5
netsnmp-python==1.0a1
nftables==0.1
numpy==1.22.3
oauthlib==3.2.0
onboard==1.4.1
opencv-python==4.5.5.64
opt-einsum==3.3.0
optimus-manager==1.4
ordered-set==4.0.2
packaging==21.3
pandas==1.4.2
parso==0.8.3
pathtools==0.1.2
peewee==3.14.10
pep517==0.12.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
ply==3.11
progress==1.6
promise==2.3
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pwquality==1.4.4
py3nvml==0.2.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.21.0
pycparser==2.21
pycups==2.0.1
pycurl==7.45.1
Pygments==2.12.0
PyGObject==3.42.1
pyOpenSSL==22.0.0
pyparsing==3.0.8
pyparted==3.11.7
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
python-dateutil==2.8.2
python-xapp==2.2.1
python-xlib==0.31
pytz==2022.1
pyxdg==0.27
PyYAML==6.0
pyzmq==22.3.0
ranger-fm==1.9.3
Reflector==2021.11.20.2.41.3
requests==2.27.1
requests-oauthlib==1.3.1
resolvelib==0.5.5
retrying==1.3.3
rsa==4.8
Rx==3.2.0
scipy==1.8.0
seaborn==0.11.2
sentry-sdk==1.5.11
setproctitle==1.2.3
shortuuid==1.0.8
six==1.16.0
smbus==1.1
smmap==5.0.0
soupsieve==2.3.1
stack-data==0.2.0
team==1.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.25.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
Theano==1.0.5
toml==0.10.2
tomli==2.0.1
torch==1.11.0
torch-tb-profiler==0.4.0
torchaudio==0.11.0
torchvision==0.12.0
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
typing_extensions==4.2.0
udiskie==2.4.2
urllib3==1.26.9
validate==5.1.0.dev0
variety==0.8.7
wandb==0.12.16
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.2
wrapt==1.14.1
youtube-dl==2021.12.17
zipp==3.8.0
